Reorganize bucket
