+++ radix-engine/src/engine/native_interpreter.rs
@@ -118,8 +118,7 @@ impl NativeInterpreter {
     {
         match (receiver.clone(), native_method.clone()) {
             (Receiver::Consumed(node_id), NativeMethod::Bucket(method)) => {
-                BucketSubstate::consuming_main(node_id, method, input, system_api)
-                    .map_err(|e| e.into())
+                Bucket::consuming_main(node_id, method, input, system_api).map_err(|e| e.into())
             }
             (Receiver::Consumed(node_id), NativeMethod::Proof(method)) => {
                 Proof::main_consume(node_id, method, input, system_api).map_err(|e| e.into())
@@ -129,7 +128,7 @@ impl NativeInterpreter {
                     .map_err(|e| e.into())
             }
             (Receiver::Ref(RENodeId::Bucket(bucket_id)), NativeMethod::Bucket(method)) => {
-                BucketSubstate::main(bucket_id, method, input, system_api).map_err(|e| e.into())
+                Bucket::main(bucket_id, method, input, system_api).map_err(|e| e.into())
             }
             (Receiver::Ref(RENodeId::Proof(proof_id)), NativeMethod::Proof(method)) => {
                 Proof::main(proof_id, method, input, system_api).map_err(|e| e.into())
